Conflicts:
	app/controllers/AbsenceRequests.java
	app/helpers/TemplateUtility.java
	app/manager/NotificationManager.java
	app/manager/configurations/EpasParam.java
	app/manager/flows/AbsenceRequestManager.java
	app/views/_workflowsMenu.html
	conf/messages.it
	conf/permissions.drl
	db/evolutions/142.sql

+++ app/controllers/Groups.java
@@ -1,20 +1,30 @@
 package controllers;
 
 import com.google.common.base.Optional;
+import dao.GeneralSettingDao;
 import dao.GroupDao;
 import dao.OfficeDao;
 import dao.PersonDao;
+import dao.RoleDao;
+import dao.UsersRolesOfficesDao;
 import helpers.Web;
+import helpers.jpa.JpaReferenceBinder;
+import java.time.LocalDate;
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 import javax.inject.Inject;
 import lombok.extern.slf4j.Slf4j;
+import lombok.val;
 import manager.GroupManager;
+import models.GeneralSetting;
 import models.Office;
 import models.Person;
 import models.Role;
 import models.User;
 import models.UsersRolesOffices;
 import models.flows.Group;
+import play.data.binding.As;
 import play.data.validation.Valid;
 import play.data.validation.Validation;
 import play.mvc.Controller;
@@ -35,27 +45,43 @@ public class Groups extends Controller {
   private static GroupManager groupManager;
   @Inject
   private static PersonDao personDao;
+  @Inject
+  private static GeneralSettingDao settingDao;
+  @Inject
+  private static UsersRolesOfficesDao uroDao;
+  @Inject
+  private static RoleDao roleDao;
 
   /**
    * Metodo che crea il gruppo.
    * @param group il gruppo da creare
    * @param office la sede su cui crearlo
    */
-  public static void createGroup(@Valid Group group, Office office) {
-
+  public static void createGroup(
+      @Valid Group group, Office office,
+      @As(binder = JpaReferenceBinder.class)
+      Set<Person> people) {
+    log.info("affiltionaPeople = {}", people);
     if (Validation.hasErrors()) {
       response.status = 400;
-      render("@blank", office);
+      List<Person> peopleForGroups = personDao.byInstitute(office.institute);
+      log.info("Create groups errors = {}", validation.errorsMap());
+      render("@edit", group, office, peopleForGroups);
     }
     rules.checkIfPermitted(group.office);
     group.office = office;
+    final boolean isNew = group.id != null;        
     group.save();
-    log.debug("Salvato nuovo gruppo di lavoro: {} per la sede {}", group.name, group.office);
+    log.debug("Salvato gruppo di lavoro: {} per la sede {}", group.name, group.office);
+
+    groupManager.updatePeople(group, people);
+
     UsersRolesOffices uro = new UsersRolesOffices();
     groupManager.createManager(office, group, uro);
-    
-    flash.success("Nuovo gruppo  di lavoro %s salvato correttamente.",
-        group.name);
+    String message = isNew 
+        ? "Nuovo gruppo di lavoro %s salvato correttamente." 
+            : "Gruppo  di lavoro %s salvato correttamente.";
+    flash.success(message, group.name);
     showGroups(office.id);
   }
 
@@ -66,17 +92,23 @@ public class Groups extends Controller {
   public static void deleteGroup(long groupId) {
     final Group group = Group.findById(groupId);
     notFoundIfNull(group);
-    rules.checkIfPermitted(group.manager.office);
+    rules.checkIfPermitted(group.office);
 
     //elimino il ruolo di manager
-    if (!groupManager.deleteManager(group)) {
-      flash.error("Non esiste un manager associato al gruppo {}. "
-          + "Impossibile eliminarlo.", group.name);
-      showGroups(group.manager.office.id);
-    } 
-    //elimino il gruppo.
-    group.delete();
-    log.debug("Eliminato gruppo {}", group.name);
+    groupManager.deleteManager(group);
+
+    if (group.getPeople().isEmpty()) {
+      //Elimino eventuali vecchie associazioni
+      group.affiliations.stream().forEach(a -> a.delete());
+      //elimino il gruppo.
+      group.delete();
+      log.info("Eliminato gruppo {}", group.name);
+    } else {
+      group.endDate = LocalDate.now();
+      group.save();
+      log.info("Disattivato gruppo {}", group.name);
+    }
+
     flash.success(Web.msgDeleted(Group.class));
     showGroups(group.manager.office.id);
   }
@@ -91,17 +123,25 @@ public class Groups extends Controller {
     rules.checkIfPermitted(office);
     User user = Security.getUser().get();
     List<Group> groups = null;
-    if (user.isSystemUser()) {
-      groups = groupDao.groupsByOffice(office, Optional.<Person>absent());
+    if (uroDao.getUsersRolesOffices(user, roleDao.getRoleByName(Role.GROUP_MANAGER), office)
+        .isPresent()) {
+      groups = 
+          groupDao.groupsByOffice(office, Optional.fromNullable(user.person), Optional.of(true));
     }
-    if (user.hasRoles(Role.PERSONNEL_ADMIN)) {
-      groups = groupDao.groupsByOffice(office, Optional.<Person>absent());
+    if (user.isSystemUser() 
+        || uroDao.getUsersRolesOffices(user, roleDao.getRoleByName(Role.PERSONNEL_ADMIN), office)
+        .isPresent()) {
+      groups = groupDao.groupsByOffice(office, Optional.<Person>absent(), Optional.of(true));
     }
-    if (user.hasRoles(Role.GROUP_MANAGER)) {
-      groups = groupDao.groupsByOffice(office, Optional.fromNullable(user.person));
-    }
-     
-    render(groups, office);
+    val activeGroups = groups.stream().filter(g -> g.isActive())
+        .sorted((g1, g2) -> 
+            g1.getName().toLowerCase().compareTo(g2.getName().toLowerCase()))
+         .collect(Collectors.toList());
+    val disabledGroups = groups.stream().filter(g -> !g.isActive())
+        .sorted((g1, g2) -> 
+            g1.getName().toLowerCase().compareTo(g2.getName().toLowerCase()))
+        .collect(Collectors.toList());
+    render(activeGroups, disabledGroups, office);
   }
 
   /**
@@ -125,23 +165,14 @@ public class Groups extends Controller {
     Office office = officeDao.getOfficeById(officeId);
     notFoundIfNull(office);
     rules.checkIfPermitted(office);
-    List<Person> peopleForGroups = personDao.byInstitute(office.institute);
+    List<Person> peopleForGroups = null;
+    GeneralSetting settings = settingDao.generalSetting();
+    if (settings.handleGroupsByInstitute) {
+      peopleForGroups = personDao.byInstitute(office.institute);
+    } else {
+      peopleForGroups = personDao.byOffice(office);
+    }
     render("@edit", office, peopleForGroups);
   }
-  
-  //  public static void manageGroup() {
-  //    User currentUser = Security.getUser().get();
-  //    List<Group> managerGroups = Lists.newArrayList();
-  //    if (currentUser.isSystemUser()) {
-  //      managerGroups = groupDao.groupsByManager(Optional.<Person>absent());
-  //    }
-  //    if (!currentUser.hasRoles(Role.GROUP_MANAGER)) {
-  //      flash.error("L'utente non dispone dei diritti per accedere alla funzionalità");
-  //      Application.index();
-  //    }
-  //    //TODO: fare la regola drools per accedere alla funzionalità
-  //    rules.checkIfPermitted();
-  //    managerGroups = groupDao.groupsByManager(Optional.fromNullable(currentUser.person));
-  //    render(managerGroups);
-  //  }
+
 }
